# Tx pool reorg optimization