--- a/core/tx_list.go
+++ b/core/tx_list.go
@@ -341,7 +341,6 @@ func (l *txList) Filter(costLimit *uint256.Int, gasLimit uint64) (types.Transact
 
 	// Filter out all the transactions above the account's funds
 	removed := l.txs.Filter(func(tx *types.Transaction) bool {
-		// SUSP!
 		cost, _ := uint256.FromBig(tx.Cost())
 		return tx.Gas() > gasLimit || cost.Cmp(costLimit) > 0
 	})