It mostly keeps the same logic and just rearranges variables.

Was BenchmarkPoolMultiAccountBatchInsert-10    	   10000	    168668 ns/op	   19387 B/op	      81 allocs/op
Now BenchmarkPoolMultiAccountBatchInsert-10    	   10000	    100608 ns/op	    2895 B/op	      52 allocs/op

a bit better 'now':
BenchmarkPoolMultiAccountBatchInsert-10    	   12112	    104955 ns/op	    2678 B/op	      47 allocs/op

```
name                                      old time/op    new time/op    delta
PoolBatchInsert100-10                       5.95ms ± 3%    5.97ms ± 0%     ~     (p=0.730 n=5+4)
PoolBatchInsert1000-10                      59.3ms ± 2%    59.0ms ± 0%     ~     (p=0.690 n=5+5)
PoolBatchInsert10000-10                      591ms ± 1%     594ms ± 0%     ~     (p=0.556 n=5+4)
PoolBatchLocalInsert100-10                  6.58ms ± 1%    6.53ms ± 2%     ~     (p=0.151 n=5+5)
PoolBatchLocalInsert1000-10                 68.8ms ± 6%    65.3ms ± 5%     ~     (p=0.056 n=5+5)
PoolBatchLocalInsert10000-10                 665ms ± 3%     639ms ± 4%     ~     (p=0.056 n=5+5)
PoolMultiAccountBatchInsert-10               169µs ±10%     108µs ±16%  -35.88%  (p=0.008 n=5+5)
PoolMultiAccountBatchInsertRace-10          2.94ms ± 9%    0.43ms ±16%  -85.48%  (p=0.008 n=5+5)
PoolMultiAccountBatchInsertNoLockRace-10     10.0s ± 0%     10.0s ± 0%     ~     (p=0.310 n=5+5)


name                                      old allocs/op  new allocs/op  delta
PoolBatchInsert100-10                        5.17k ± 0%     2.99k ± 0%  -42.15%  (p=0.008 n=5+5)
PoolBatchInsert1000-10                       51.7k ± 0%     29.8k ± 0%  -42.31%  (p=0.008 n=5+5)
PoolBatchInsert10000-10                       474k ± 0%      260k ± 0%  -45.17%  (p=0.008 n=5+5)
PoolBatchLocalInsert100-10                   5.95k ± 0%     3.24k ± 0%  -45.44%  (p=0.008 n=5+5)
PoolBatchLocalInsert1000-10                  58.8k ± 0%     31.8k ± 0%  -45.92%  (p=0.008 n=5+5)
PoolBatchLocalInsert10000-10                  587k ± 0%      317k ± 0%  -45.97%  (p=0.008 n=5+5)
PoolMultiAccountBatchInsert-10                79.0 ± 0%      51.0 ± 0%  -35.44%  (p=0.000 n=5+4)
PoolMultiAccountBatchInsertRace-10           14.8k ±10%      2.3k ± 1%  -84.47%  (p=0.008 n=5+5)
PoolMultiAccountBatchInsertNoLockRace-10       148 ±20%       113 ± 8%  -23.56%  (p=0.029 n=4+4)

```

[old.txt](https://github.com/maticnetwork/bor/files/9841925/old.txt)