@JekaMas 2022-10-22
Latest update:

```
benchstat new.txt old.txt
 
name                                      old time/op    new time/op    delta
PoolBatchInsert100-10                       5.82ms ± 1%    6.09ms ± 0%    +4.76%  (p=0.008 n=5+5)
PoolBatchInsert1000-10                      57.8ms ± 2%    60.7ms ± 0%    +4.88%  (p=0.008 n=5+5)
PoolBatchInsert10000-10                      589ms ± 2%     598ms ± 0%      ~     (p=0.690 n=5+5)
PoolBatchLocalInsert100-10                  6.57ms ± 2%    6.66ms ± 1%      ~     (p=0.056 n=5+5)
PoolBatchLocalInsert1000-10                 65.2ms ± 1%    66.4ms ± 0%    +1.84%  (p=0.008 n=5+5)
PoolBatchLocalInsert10000-10                 638ms ± 1%     662ms ± 2%    +3.80%  (p=0.008 n=5+5)
PoolMultiAccountBatchInsert-10               114µs ± 2%     158µs ± 1%   +39.24%  (p=0.008 n=5+5)
PoolMultiAccountBatchInsertRace-10           487µs ± 3%    3040µs ±10%  +523.56%  (p=0.008 n=5+5)
PoolMultiAccountBatchInsertNoLockRace-10     277µs ± 1%     810µs ± 7%  +192.13%  (p=0.008 n=5+5)

name                                      old alloc/op   new alloc/op   delta
PoolBatchInsert100-10                        154kB ± 0%     204kB ± 0%   +33.00%  (p=0.008 n=5+5)
PoolBatchInsert1000-10                      1.57MB ± 0%    2.08MB ± 0%   +32.93%  (p=0.008 n=5+5)
PoolBatchInsert10000-10                     14.8MB ± 0%    19.0MB ± 0%   +27.94%  (p=0.008 n=5+5)
PoolBatchLocalInsert100-10                  4.16MB ± 3%    4.04MB ± 2%      ~     (p=0.095 n=5+5)
PoolBatchLocalInsert1000-10                 40.3MB ± 5%    39.5MB ± 0%      ~     (p=0.690 n=5+5)
PoolBatchLocalInsert10000-10                 425MB ± 0%     431MB ± 0%    +1.44%  (p=0.008 n=5+5)
PoolMultiAccountBatchInsert-10              2.91kB ± 0%   19.36kB ± 0%  +565.09%  (p=0.016 n=5+4)
PoolMultiAccountBatchInsertRace-10           430kB ± 0%    3217kB ± 7%  +647.51%  (p=0.008 n=5+5)
PoolMultiAccountBatchInsertNoLockRace-10     253kB ± 1%     863kB ± 6%  +241.74%  (p=0.008 n=5+5)

name                                      old allocs/op  new allocs/op  delta
PoolBatchInsert100-10                        2.98k ± 0%     5.17k ± 0%   +73.40%  (p=0.008 n=5+5)
PoolBatchInsert1000-10                       29.8k ± 0%     51.8k ± 0%   +73.66%  (p=0.008 n=5+5)
PoolBatchInsert10000-10                       266k ± 0%      474k ± 0%   +77.98%  (p=0.008 n=5+5)
PoolBatchLocalInsert100-10                   3.24k ± 0%     5.95k ± 0%   +83.45%  (p=0.008 n=5+5)
PoolBatchLocalInsert1000-10                  31.8k ± 0%     58.8k ± 0%   +84.93%  (p=0.008 n=5+5)
PoolBatchLocalInsert10000-10                  317k ± 0%      587k ± 0%   +85.10%  (p=0.008 n=5+5)
PoolMultiAccountBatchInsert-10                50.0 ± 0%      79.0 ± 0%   +58.00%  (p=0.008 n=5+5)
PoolMultiAccountBatchInsertRace-10           2.46k ± 1%    16.65k ± 9%  +577.10%  (p=0.008 n=5+5)
PoolMultiAccountBatchInsertNoLockRace-10     1.32k ± 1%     4.43k ± 6%  +234.80%  (p=0.008 n=5+5)
```

[new.txt](https://github.com/maticnetwork/bor/files/9845093/new.txt)
[old.txt](https://github.com/maticnetwork/bor/files/9845094/old.txt)