@JekaMas 2022-10-23
```
benchstat new_locals_err.txt new_locals.txt 
name                                      old time/op    new time/op    delta
PoolBatchInsert100-10                       5.81ms ± 2%    6.01ms ± 1%     +3.45%  (p=0.008 n=5+5)
PoolBatchInsert1000-10                      58.1ms ± 3%    60.0ms ± 1%       ~     (p=0.056 n=5+5)
PoolBatchInsert10000-10                      627ms ±15%     594ms ± 0%       ~     (p=0.151 n=5+5)
PoolBatchLocalInsert100-10                  6.52ms ± 2%    6.51ms ± 0%       ~     (p=0.841 n=5+5)
PoolBatchLocalInsert1000-10                 64.4ms ± 2%    65.3ms ± 1%       ~     (p=0.151 n=5+5)
PoolBatchLocalInsert10000-10                 637ms ± 3%     639ms ± 0%       ~     (p=0.190 n=5+4)
PoolAccountMultiBatchInsert-10               114µs ± 1%     114µs ± 1%       ~     (p=0.222 n=5+5)
PoolAccountMultiBatchInsertRace-10           433µs ± 5%     545µs ± 7%    +25.81%  (p=0.008 n=5+5)
PoolAccountMultiBatchInsertNoLockRace-10     245µs ± 9%     308µs ± 2%    +25.72%  (p=0.008 n=5+5)
PoolAccountsBatchInsert-10                   116µs ± 9%     115µs ± 1%       ~     (p=1.000 n=5+5)
PoolAccountsBatchInsertRace-10               433µs ±13%     496µs ± 2%    +14.72%  (p=0.016 n=5+5)
PoolAccountsBatchInsertNoLockRace-10         262µs ±10%     302µs ± 9%    +15.07%  (p=0.016 n=5+5)

name                                      old alloc/op   new alloc/op   delta
PoolBatchInsert100-10                        153kB ± 0%     158kB ± 0%     +2.69%  (p=0.008 n=5+5)
PoolBatchInsert1000-10                      1.56MB ± 0%    1.61MB ± 0%     +2.93%  (p=0.016 n=4+5)
PoolBatchInsert10000-10                     14.7MB ± 0%    14.5MB ± 0%     -0.82%  (p=0.008 n=5+5)
PoolBatchLocalInsert100-10                  4.16MB ± 2%    4.13MB ± 1%       ~     (p=0.841 n=5+5)
PoolBatchLocalInsert1000-10                 40.3MB ± 5%    38.9MB ± 0%       ~     (p=0.690 n=5+5)
PoolBatchLocalInsert10000-10                 425MB ± 0%     426MB ± 0%     +0.10%  (p=0.008 n=5+5)
PoolAccountMultiBatchInsert-10              2.90kB ± 0%    2.93kB ± 0%     +1.08%  (p=0.008 n=5+5)
PoolAccountMultiBatchInsertRace-10           424kB ± 2%     430kB ± 2%       ~     (p=0.095 n=5+5)
PoolAccountMultiBatchInsertNoLockRace-10     242kB ± 0%     255kB ± 0%     +5.65%  (p=0.008 n=5+5)
PoolAccountsBatchInsert-10                  2.88kB ± 0%    2.93kB ± 0%     +1.58%  (p=0.008 n=5+5)
PoolAccountsBatchInsertRace-10               414kB ± 6%     409kB ± 0%       ~     (p=0.690 n=5+5)
PoolAccountsBatchInsertNoLockRace-10         241kB ± 2%     255kB ± 0%     +5.96%  (p=0.016 n=5+4)

name                                      old allocs/op  new allocs/op  delta
PoolBatchInsert100-10                        2.98k ± 0%     2.99k ± 0%       ~     (p=0.063 n=5+5)
PoolBatchInsert1000-10                       29.8k ± 0%     29.9k ± 0%     +0.32%  (p=0.016 n=4+5)
PoolBatchInsert10000-10                       260k ± 0%      260k ± 0%       ~     (p=0.056 n=5+5)
PoolBatchLocalInsert100-10                   3.24k ± 0%     3.25k ± 0%     +0.10%  (p=0.016 n=5+4)
PoolBatchLocalInsert1000-10                  31.8k ± 0%     31.8k ± 0%       ~     (p=0.135 n=5+5)
PoolBatchLocalInsert10000-10                  317k ± 0%      317k ± 0%       ~     (p=0.389 n=5+5)
PoolAccountMultiBatchInsert-10                49.0 ± 0%      51.0 ± 0%     +4.08%  (p=0.008 n=5+5)
PoolAccountMultiBatchInsertRace-10             118 ± 1%      2459 ± 3%  +1980.20%  (p=0.008 n=5+5)
PoolAccountMultiBatchInsertNoLockRace-10      82.0 ± 0%    1338.6 ± 0%  +1532.44%  (p=0.008 n=5+5)
PoolAccountsBatchInsert-10                    47.0 ± 0%      51.0 ± 0%     +8.51%  (p=0.008 n=5+5)
PoolAccountsBatchInsertRace-10                 115 ± 3%      2318 ± 0%  +1912.33%  (p=0.008 n=5+5)
PoolAccountsBatchInsertNoLockRace-10          81.0 ± 0%    1339.2 ± 0%  +1553.40%  (p=0.029 n=4+4)
jekamas@MacBook-Pro-Evgeniy bor % benchstat new_locals.txt new_locals_err.txt 
name                                      old time/op    new time/op    delta
PoolBatchInsert100-10                       6.01ms ± 1%    5.81ms ± 2%   -3.33%  (p=0.008 n=5+5)
PoolBatchInsert1000-10                      60.0ms ± 1%    58.1ms ± 3%     ~     (p=0.056 n=5+5)
PoolBatchInsert10000-10                      594ms ± 0%     627ms ±15%     ~     (p=0.151 n=5+5)
PoolBatchLocalInsert100-10                  6.51ms ± 0%    6.52ms ± 2%     ~     (p=0.841 n=5+5)
PoolBatchLocalInsert1000-10                 65.3ms ± 1%    64.4ms ± 2%     ~     (p=0.151 n=5+5)
PoolBatchLocalInsert10000-10                 639ms ± 0%     637ms ± 3%     ~     (p=0.190 n=4+5)
PoolAccountMultiBatchInsert-10               114µs ± 1%     114µs ± 1%     ~     (p=0.222 n=5+5)
PoolAccountMultiBatchInsertRace-10           545µs ± 7%     433µs ± 5%  -20.52%  (p=0.008 n=5+5)
PoolAccountMultiBatchInsertNoLockRace-10     308µs ± 2%     245µs ± 9%  -20.46%  (p=0.008 n=5+5)
PoolAccountsBatchInsert-10                   115µs ± 1%     116µs ± 9%     ~     (p=1.000 n=5+5)
PoolAccountsBatchInsertRace-10               496µs ± 2%     433µs ±13%  -12.83%  (p=0.016 n=5+5)
PoolAccountsBatchInsertNoLockRace-10         302µs ± 9%     262µs ±10%  -13.10%  (p=0.016 n=5+5)

name                                      old alloc/op   new alloc/op   delta
PoolBatchInsert100-10                        158kB ± 0%     153kB ± 0%   -2.62%  (p=0.008 n=5+5)
PoolBatchInsert1000-10                      1.61MB ± 0%    1.56MB ± 0%   -2.85%  (p=0.016 n=5+4)
PoolBatchInsert10000-10                     14.5MB ± 0%    14.7MB ± 0%   +0.83%  (p=0.008 n=5+5)
PoolBatchLocalInsert100-10                  4.13MB ± 1%    4.16MB ± 2%     ~     (p=0.841 n=5+5)
PoolBatchLocalInsert1000-10                 38.9MB ± 0%    40.3MB ± 5%     ~     (p=0.690 n=5+5)
PoolBatchLocalInsert10000-10                 426MB ± 0%     425MB ± 0%   -0.10%  (p=0.008 n=5+5)
PoolAccountMultiBatchInsert-10              2.93kB ± 0%    2.90kB ± 0%   -1.07%  (p=0.008 n=5+5)
PoolAccountMultiBatchInsertRace-10           430kB ± 2%     424kB ± 2%     ~     (p=0.095 n=5+5)
PoolAccountMultiBatchInsertNoLockRace-10     255kB ± 0%     242kB ± 0%   -5.35%  (p=0.008 n=5+5)
PoolAccountsBatchInsert-10                  2.93kB ± 0%    2.88kB ± 0%   -1.56%  (p=0.008 n=5+5)
PoolAccountsBatchInsertRace-10               409kB ± 0%     414kB ± 6%     ~     (p=0.690 n=5+5)
PoolAccountsBatchInsertNoLockRace-10         255kB ± 0%     241kB ± 2%   -5.63%  (p=0.016 n=4+5)

name                                      old allocs/op  new allocs/op  delta
PoolBatchInsert100-10                        2.99k ± 0%     2.98k ± 0%     ~     (p=0.063 n=5+5)
PoolBatchInsert1000-10                       29.9k ± 0%     29.8k ± 0%   -0.32%  (p=0.016 n=5+4)
PoolBatchInsert10000-10                       260k ± 0%      260k ± 0%     ~     (p=0.056 n=5+5)
PoolBatchLocalInsert100-10                   3.25k ± 0%     3.24k ± 0%   -0.10%  (p=0.000 n=4+5)
PoolBatchLocalInsert1000-10                  31.8k ± 0%     31.8k ± 0%     ~     (p=0.135 n=5+5)
PoolBatchLocalInsert10000-10                  317k ± 0%      317k ± 0%     ~     (p=0.389 n=5+5)
PoolAccountMultiBatchInsert-10                51.0 ± 0%      49.0 ± 0%   -3.92%  (p=0.008 n=5+5)
PoolAccountMultiBatchInsertRace-10           2.46k ± 3%     0.12k ± 1%  -95.19%  (p=0.008 n=5+5)
PoolAccountMultiBatchInsertNoLockRace-10     1.34k ± 0%     0.08k ± 0%  -93.87%  (p=0.008 n=5+5)
PoolAccountsBatchInsert-10                    51.0 ± 0%      47.0 ± 0%   -7.84%  (p=0.008 n=5+5)
PoolAccountsBatchInsertRace-10               2.32k ± 0%     0.12k ± 3%  -95.03%  (p=0.008 n=5+5)
PoolAccountsBatchInsertNoLockRace-10         1.34k ± 0%     0.08k ± 0%  -93.95%  (p=0.029 n=4+4)

```
[new_locals_err.txt](https://github.com/maticnetwork/bor/files/9846413/new_locals_err.txt)
[new_locals.txt](https://github.com/maticnetwork/bor/files/9846414/new_locals.txt)