@JekaMas 2022-10-23
and the same comparing to develop branch:

```
benchstat old_locals.txt new_locals_err.txt                
name                                      old time/op    new time/op    delta
PoolBatchInsert100-10                       6.06ms ± 2%    5.81ms ± 2%   -4.11%  (p=0.008 n=5+5)
PoolBatchInsert1000-10                      60.0ms ± 0%    58.1ms ± 3%     ~     (p=0.056 n=5+5)
PoolBatchInsert10000-10                      591ms ± 0%     627ms ±15%     ~     (p=0.190 n=4+5)
PoolBatchLocalInsert100-10                  6.56ms ± 1%    6.52ms ± 2%     ~     (p=0.310 n=5+5)
PoolBatchLocalInsert1000-10                 66.8ms ±14%    64.4ms ± 2%     ~     (p=0.548 n=5+5)
PoolBatchLocalInsert10000-10                 670ms ± 6%     637ms ± 3%   -4.84%  (p=0.032 n=5+5)
PoolAccountMultiBatchInsert-10               164µs ± 3%     114µs ± 1%  -30.63%  (p=0.008 n=5+5)
PoolAccountMultiBatchInsertRace-10          2.92ms ± 6%    0.43ms ± 5%  -85.14%  (p=0.008 n=5+5)
PoolAccountMultiBatchInsertNoLockRace-10     816µs ± 6%     245µs ± 9%  -69.93%  (p=0.008 n=5+5)
PoolAccountsBatchInsert-10                   157µs ± 1%     116µs ± 9%  -26.35%  (p=0.016 n=4+5)
PoolAccountsBatchInsertRace-10              2.03ms ± 5%    0.43ms ±13%  -78.71%  (p=0.008 n=5+5)
PoolAccountsBatchInsertNoLockRace-10         837µs ± 4%     262µs ±10%  -68.63%  (p=0.008 n=5+5)

name                                      old alloc/op   new alloc/op   delta
PoolBatchInsert100-10                        204kB ± 0%     153kB ± 0%  -24.89%  (p=0.008 n=5+5)
PoolBatchInsert1000-10                      2.08MB ± 0%    1.56MB ± 0%  -25.10%  (p=0.016 n=5+4)
PoolBatchInsert10000-10                     19.0MB ± 0%    14.7MB ± 0%  -22.76%  (p=0.008 n=5+5)
PoolBatchLocalInsert100-10                  4.10MB ± 1%    4.16MB ± 2%     ~     (p=0.056 n=5+5)
PoolBatchLocalInsert1000-10                 38.0MB ± 6%    40.3MB ± 5%     ~     (p=0.548 n=5+5)
PoolBatchLocalInsert10000-10                 431MB ± 0%     425MB ± 0%   -1.44%  (p=0.008 n=5+5)
PoolAccountMultiBatchInsert-10              19.4kB ± 0%     2.9kB ± 0%  -85.04%  (p=0.008 n=5+5)
PoolAccountMultiBatchInsertRace-10          3.16MB ± 6%    0.42MB ± 2%  -86.59%  (p=0.008 n=5+5)
PoolAccountMultiBatchInsertNoLockRace-10     803kB ± 6%     242kB ± 0%  -69.92%  (p=0.008 n=5+5)
PoolAccountsBatchInsert-10                  19.4kB ± 0%     2.9kB ± 0%  -85.11%  (p=0.008 n=5+5)
PoolAccountsBatchInsertRace-10              1.78MB ± 3%    0.41MB ± 6%  -76.77%  (p=0.008 n=5+5)
PoolAccountsBatchInsertNoLockRace-10         832kB ± 3%     241kB ± 2%  -71.03%  (p=0.008 n=5+5)

name                                      old allocs/op  new allocs/op  delta
PoolBatchInsert100-10                        5.17k ± 0%     2.98k ± 0%  -42.29%  (p=0.016 n=4+5)
PoolBatchInsert1000-10                       51.8k ± 0%     29.8k ± 0%  -42.49%  (p=0.029 n=4+4)
PoolBatchInsert10000-10                       474k ± 0%      260k ± 0%  -45.17%  (p=0.008 n=5+5)
PoolBatchLocalInsert100-10                   5.95k ± 0%     3.24k ± 0%  -45.51%  (p=0.000 n=4+5)
PoolBatchLocalInsert1000-10                  58.8k ± 0%     31.8k ± 0%  -45.92%  (p=0.008 n=5+5)
PoolBatchLocalInsert10000-10                  587k ± 0%      317k ± 0%  -45.97%  (p=0.008 n=5+5)
PoolAccountMultiBatchInsert-10                79.0 ± 0%      49.0 ± 0%  -37.97%  (p=0.008 n=5+5)
PoolAccountMultiBatchInsertRace-10           16.4k ± 6%      0.1k ± 1%  -99.28%  (p=0.008 n=5+5)
PoolAccountMultiBatchInsertNoLockRace-10     4.09k ± 7%     0.08k ± 0%  -97.99%  (p=0.008 n=5+5)
PoolAccountsBatchInsert-10                    79.0 ± 0%      47.0 ± 0%  -40.51%  (p=0.008 n=5+5)
PoolAccountsBatchInsertRace-10               10.1k ± 4%      0.1k ± 3%  -98.86%  (p=0.008 n=5+5)
PoolAccountsBatchInsertNoLockRace-10         4.27k ± 3%     0.08k ± 0%  -98.10%  (p=0.016 n=5+4)

```
[new_locals_err.txt](https://github.com/maticnetwork/bor/files/9846428/new_locals_err.txt)
[old_locals.txt](https://github.com/maticnetwork/bor/files/9846429/old_locals.txt)